reorg: moved config file into separate folder; renamed get_surface_code_reg_fn to get_hamiltonian_reg_fn; style fixes.
